week2/fastapi_service/requirements.txt
fastapi==0.111.0
numpy==2.0.0
pydantic==2.7.4
python-dotenv==1.0.1
word-vectors==4.0.2